+++ AdventOfCode2023/AdventOfCode2023/ISolveProblems.cs
@@ -2,5 +2,5 @@ namespace AdventOfCode2023;
 
 public interface ISolveProblems
 {
-    public int Solve(string[] lines);
+    public long Solve(string[] lines);
 }
